# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Goods Rental and Sale API",
        "description": "API service for renting or selling goods. Users can purchase or rent products, renew rentals, and check product status.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/products/{productId}/purchase": {
            "post": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Purchase a product permanently",
                "operationId": "1ac038c9a43edbec6d77432ecbc65164",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of the product to purchase",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product purchased successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserProduct"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (e.g., insufficient balance)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/products/{productId}/rent": {
            "post": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Rent a product for a specified duration",
                "operationId": "d8fc19cfe814cb1a810c52524a52e158",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of the product to rent",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rental successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product rented successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserProduct"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (e.g., insufficient balance, duration exceeds limit)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/products/{productId}/renew": {
            "post": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Renew an existing rental",
                "operationId": "4cf2b7b866c1de45c4986dc59edd38a7",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of the product to renew rental for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RenewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rental renewal successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rental renewed successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserProduct"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (e.g., duration exceeds total limit)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (e.g., no active rental found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/user/purchase-history": {
            "get": {
                "tags": [
                    "User Activities"
                ],
                "summary": "Get user's purchase and rental history",
                "operationId": "b02650913e044cf57f3974685b51aaa0",
                "responses": {
                    "200": {
                        "description": "User purchase and rental history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserProduct"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
...
            "UserProduct": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1001
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 101
                    },
                    "ownership_type": {
                        "type": "string",
                        "example": "rent"
                    },
                    "rent_started_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-10T15:00:00.000Z"
                    },
                    "rent_expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-10T19:00:00.000Z"
                    },
                    "unique_code": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-10T14:50:00.000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-10T14:50:00.000Z"
                    }
                },
                "type": "object"
            },
            "RentRequest": {
                "required": [
                    "duration"
                ],
                "properties": {
                    "duration": {
                        "type": "integer",
                        "enum": [
                            4,
                            8,
                            12,
                            24
                        ],
                        "example": 8
                    }
                },
                "type": "object"
            },
            "RenewRequest": {
                "required": [
                    "duration"
                ],
                "properties": {
                    "duration": {
                        "type": "integer",
                        "enum": [
                            4,
                            8,
                            12,
                            24
                        ],
                        "example": 4
                    }
                },
                "type": "object"
            }
        }
    },
...
}